Add RecordFoodService Logic
