+++ src/main/java/com/prob_jr/sikcal_app/domain/service/RecordFoodService.java
@@ -4,10 +4,7 @@ import com.prob_jr.sikcal_app.domain.Food;
 import com.prob_jr.sikcal_app.domain.Member;
 import com.prob_jr.sikcal_app.domain.Record;
 import com.prob_jr.sikcal_app.domain.RecordFood;
-import com.prob_jr.sikcal_app.domain.repository.FoodRepository;
-import com.prob_jr.sikcal_app.domain.repository.MemberRepository;
-import com.prob_jr.sikcal_app.domain.repository.RecordFoodRepository;
-import com.prob_jr.sikcal_app.domain.repository.RecordRepository;
+import com.prob_jr.sikcal_app.domain.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -45,17 +42,17 @@ public class RecordFoodService {
      * 식단 삭제
      */
     @Transactional
-    public void cancelRecord(Long recordId) {
+    public void deleteRecord(Long recordId) {
         //엔티티 조회
         List<RecordFood> foods = recordFoodRepository.findFoods(recordId);
         Record record = recordRepository.findOne(recordId);
 
         //하나씩 recordfood 전부 삭제
         for (RecordFood recordFood : foods) {
-            foods.remove(recordFood);
+            recordFoodRepository.delete(recordFood);
         }
         //record 삭제
-        recordRepository.cancel(record);
+        recordRepository.delete(record);
     }
 
 
@@ -80,10 +77,26 @@ public class RecordFoodService {
     /**
      * 음식 삭제
      */
+    @Transactional
+    public void deleteFood(Long recordId) {
+        //Record 조회
+        Record record = recordRepository.findOne(recordId);
+        //RecordFood조회
+        RecordFood recordFood = recordFoodRepository.findOne(recordId);
+        //RecordFood삭제
+        recordFoodRepository.delete(recordFood);
 
+    }
 
     /**
      * 음식 검색
      */
+//    public List<RecordFood> search(FoodSearch foodSearch) {
+//
+//
+//    }
+
+
+
 
 }
